requirements.txt
torch~=2.0.0+cu118
numpy~=1.26.2
matplotlib~=3.6.2
pandas~=2.1.3
argparse~=1.4.0
seaborn~=0.13.0
scikit-learn~=1.4.2
sbibm~=1.1.0
qmcpy~=1.4
scipy~=1.11.3
PyYAML~=6.0.1
tqdm~=4.65.0